+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `Target` enum to better handle output printing.
 - Added `#![forbid(unsafe_code)]` to forbid unsafe code.
+- Added colored output.
 
 ### Fixed